requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
Deprecated==1.2.18
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
limits==5.2.0
packaging==25.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
wrapt==1.17.2
